comment display bug fix, starting function migration, voting bug exception

+++ app/ieddit.py
@@ -4,7 +4,10 @@ Main ieddit code.
 TODO: split this up into different views, function groups, etc.
 """
 from share import *
+from functions.db_functions import *
+from functions.functions import *
 import _thread
+
 
 @app.before_request
 def before_request():
@@ -121,14 +124,16 @@ def handle_error(error):
         description = "A(n) {} occurred. The developers have been notified of this."\
         .format(type(error).__name__  or 'unknown error')
 
-        logger.error(error)
-        logger.error(traceback.format_exc())
+        logger.error(title)
+        logger.error(str(trace_back))
         if config.DISCORD_ENABLED:
             send_discord_msg(title=title, description=str(trace_back))
 
+    print(title)
+    print(trace_back)
     return render_template("error.html", error=description, code=code), code
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_style(sub=None):
     """
     returns sub style for a given sub
@@ -213,7 +218,7 @@ def sitemap():
 def robotstxt():
     return app.send_static_file('robots.txt')
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_subtitle(sub):
     try:
         title = db.session.query(Sub).filter_by(name=sub).first()
@@ -222,12 +227,12 @@ def get_subtitle(sub):
         title = None
     return title
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_api_key(username):
     key = db.session.query(Api_key).filter_by(username=username).first()
     return key
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def has_messages(username):
     if 'username' in session:
         messages = db.session.query(Message).filter_by(sent_to=username, read=False).count()
@@ -238,7 +243,7 @@ def has_messages(username):
                 return True
     return False
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_sub_mods(sub, admin=True):
     mod_subs = db.session.query(Moderator).filter_by(sub=sub).all()
     if admin == False:
@@ -248,7 +253,7 @@ def get_sub_mods(sub, admin=True):
         mod_subs.append(a)
     return [m.username for m in mod_subs]
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_banned_subs(username):
     subs = db.session.query(Ban).filter_by(username=username).all()
     b = []
@@ -256,7 +261,7 @@ def get_banned_subs(username):
         b.append(s.sub)
     return b
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def anon_block(obj):
     for c in obj:
         if c.anonymous:
@@ -272,11 +277,11 @@ def anon_block(obj):
                 c.noblock = True
     return obj
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_muted_subs():
     return [x.name for x in db.session.query(Sub).filter_by(muted=True).all()]
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_blocked(username):
     bdict = {'comment_id':[], 'post_id':[], 'other_user':[], 'anon_user':[]}
     if username == None:
@@ -305,7 +310,7 @@ def get_blocked(username):
 
     return bdict
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def is_mod(obj, username):
     if hasattr(obj, 'inurl_title'):
         post = obj
@@ -318,7 +323,7 @@ def is_mod(obj, username):
             return True
     return False
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def is_admin(username):
     if db.session.query(db.session.query(Iuser).filter_by(admin=True, username=username).exists()).scalar():
     #if 'admin' in session:
@@ -333,7 +338,7 @@ def set_rate_limit(limit_seconds=None):
             session['rate_limit'] = int(time.time()) + (limit_seconds)
         #cache.clear()
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def normalize_username(username, dbuser=False):
     if username == None:
         return False
@@ -344,14 +349,7 @@ def normalize_username(username, dbuser=False):
         return username.username
     return False
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
-def normalize_sub(sub):
-    subl = db.session.query(Sub).filter(func.lower(Sub.name) == func.lower(sub)).first()
-    if subl != None:
-        return subl.name
-    return sub
-
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_all_subs(explore=False):
     subs = db.session.query(Sub).all()
     if explore == False:
@@ -383,7 +381,7 @@ def get_all_subs(explore=False):
         return esubs
 
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_pgp_from_username(username):
     u = normalize_username(username)
     if u == False:
@@ -396,13 +394,13 @@ def get_pgp_from_username(username):
         return pgp
     return False
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_user_from_name(username):
     if username == '' or username == False or username == None:
         return False
     return normalize_username(username, dbuser=True)
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_user_from_id(uid):
     if uid == None or uid == False:
         return False
@@ -520,18 +518,12 @@ def register():
         return redirect(config.URL, 302)
 
 @app.route('/')
-
 def index():
     return subi(subi='all', nsfw=False)
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
-def is_sub_nsfw(sub):
-    s = db.session.query(Sub).filter_by(name=sub).first()
-    if s.nsfw:
-        return True
-    return False
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_subi(subi, user_id=None, posts_only=False, deleted=False, offset=0, limit=15, nsfw=True, d=None, s=None):
     if offset != None:
         offset = int(offset)
@@ -733,11 +725,11 @@ def subi(subi, user_id=None, posts_only=False, offset=0, limit=15, nsfw=True, sh
     return render_template('sub.html', posts=sub_posts, url=config.URL, sub_stats=sub_stats, is_all=is_all)
 
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_cached_children(comment, deleted=False):
     return comment.get_children(deleted=deleted)
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def recursive_children(comment=None, current_depth=0, max_depth=8, deleted=False):
     found_children = []
     found_children.append(comment)
@@ -758,7 +750,7 @@ def recursive_children(comment=None, current_depth=0, max_depth=8, deleted=False
 
     return found_children
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def c_get_comments(sub=None, post_id=None, inurl_title=None, comment_id=False, sort_by=None, comments_only=False, user_id=None, deleted=False):
     post = None
     parent_comment = None
@@ -904,7 +896,7 @@ def get_comments(sub=None, post_id=None, inurl_title=None, comment_id=None, sort
 # this sort of recursion KILLS performance, especially when combined with the already
 # terrible comment_structure function.
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def list_of_child_comments(comment_id, sort_by=None):
     comments = {}
     current_comments = []
@@ -1097,19 +1089,15 @@ def vote(post_id=None, comment_id=None, vote=None, user_id=None):
                 elif last_vote.comment_id is not None:
                     vpost = db.session.query(Comment).filter_by(id=last_vote.comment_id).first()
 
-            try:
-                score = str(vpost.ups - vpost.downs)
-            except:
-                logger.error('VPOST is undefined, line 1103. Need to investigate why.')
-                logger.error(str(sqla_to_dict(last_vote)))
-                post = db.session.query(Post).filter_by(id=last_vote.post_id).first()
-                score = post.ups - post.downs
+                if last_vote.vote == 1:
+                    vpost.ups -= 1
+                elif last_vote.vote == -1:
+                    vpost.downs -= 1
 
+            if vpost is None:
+                raise ValueError('vpost is none. %s' % str(last_vote))
 
-            if last_vote.vote == 1:
-                vpost.ups -= 1
-            elif last_vote.vote == -1:
-                vpost.downs -= 1
+            score = str(vpost.ups - vpost.downs)
 
             db.session.delete(last_vote)
             db.session.commit()
@@ -1682,7 +1670,7 @@ def settings(sub=None):
         return render_template('sub_mods.html', mods=get_sub_mods(sub, admin=False), settings=True, nsfw=subr.nsfw, sub_object=subr)
     return abort(403)
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_blocked_subs(username=None):
     subs = db.session.query(Sub_block).filter_by(username=session['username']).all()
     if subs != None:
@@ -1724,7 +1712,7 @@ def blocksub(sub=None):
 
     return redirect('/i/%s/' % sub)
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def explore_stats(sub):
     if hasattr(sub, 'rules'):
         if sub.rules != None:
@@ -1808,12 +1796,16 @@ def subcomments(sub=None, offset=0, limit=15, s=None, nsfw=False):
 
     comments_with_posts = []
 
-    if sub == None:
+    if sub is None:
+        nsfw = False
         sub = 'all'
 
     if sub == 'all':
-        comments = db.session.query(Comment)
-        comcount = db.session.query(Comment).count()
+        if nsfw:
+            comments = db.session.query(Comment)
+        else:
+            comments = db.session.query(Comment).filter_by(nsfw=nsfw)
+        comcount = comments.count()
     else:
         comments = db.session.query(Comment).filter_by(sub_name=normalize_sub(sub))
         comcount = comments.count()
@@ -1921,15 +1913,15 @@ def subcomments(sub=None, offset=0, limit=15, s=None, nsfw=False):
 
     return render_template('recentcomments.html', posts=posts, url=config.URL, comments_with_posts=comments_with_posts, no_posts=True)
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_posts_and_comments(subi=None, day=False, load=None):
     filter_today = datetime.now() - timedelta(days=1)
     if subi == None or subi == 'all':
         posts = db.session.query(Post)
         comments = db.session.query(Comment)
     else:
-        posts = db.session.query(Post).filter_by(sub=subi)
-        comments = db.session.query(Comment).filter_by(sub_name=subi)
+        posts = db.session.query(Post)
+        comments = db.session.query(Comment)
 
     if day:
         posts = posts.filter(Post.created >= filter_today)
@@ -1947,7 +1939,7 @@ def get_posts_and_comments(subi=None, day=False, load=None):
 
     return posts, comments
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_top_stats(subi=None):
     t = time.time()
     votes = []
@@ -1975,7 +1967,7 @@ def get_top_stats(subi=None):
 
 
 
-@cache.memoize(config.DEFAULT_CACHE_TIME, unless=only_cache_get)
+@cache.memoize(config.DEFAULT_CACHE_TIME)
 def get_stats(subi=None):
     if subi == None:
         votes = db.session.query(Vote).all()
